+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Offramp: log error on service error
+- Offramp: throw RuntimeException after interrupt
+
 ## [6.0.11] - 2019-09-02
 ### Changed
 - Disabled LDAP health check.